Move metrics_transmitter binary into metrics dir
Bug: 550013241
