+++ base/cvd/cuttlefish/metrics/metrics_transmitter/BUILD.bazel
@@ -75,10 +75,10 @@ cf_cc_binary(
     ],
     deps = [
         "//cuttlefish/common/libs/utils:tee_logging",
-        "//cuttlefish/host/commands/metrics_transmitter:debug_reader",
-        "//cuttlefish/host/commands/metrics_transmitter:metrics_conversion",
-        "//cuttlefish/host/commands/metrics_transmitter:metrics_flags",
-        "//cuttlefish/host/commands/metrics_transmitter:metrics_transmission",
+        "//cuttlefish/metrics/metrics_transmitter:debug_reader",
+        "//cuttlefish/metrics/metrics_transmitter:metrics_conversion",
+        "//cuttlefish/metrics/metrics_transmitter:metrics_flags",
+        "//cuttlefish/metrics/metrics_transmitter:metrics_transmission",
         "//cuttlefish/result",
         "//external_proto:clientanalytics_cc_proto",
         "@abseil-cpp//absl/log",
