+++ base/cvd/cuttlefish/metrics/metrics_transmitter/debug_reader.h
